refactor: extract findCurrentUserInfo
