+++ src/types/UserLS.ts
@@ -1,7 +1,9 @@
+import { HistoryInfo } from './HistoryInfo';
+
 export type UserLS = {
   email: string;
   password: string;
   userID: string;
   favorites: string[];
-  history: string[];
+  history: HistoryInfo[];
 };
